apps/api/pyproject.toml
[project]
name = "course-agent-backend"
version = "0.1.0"
description = "Course generation agent backend built with FastAPI, LangGraph, and LangChain."
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "sse-starlette>=2.1.3,<3.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.7.0,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "langchain>=0.3.13,<2.0.0",
  "langchain-openai>=0.2.14,<2.0.0",
  "langgraph>=0.2.60,<2.0.0",
  "langgraph-checkpoint-sqlite>=3.0.3,<4.0.0",
  "deepagents>=0.0.8,<1.0.0",
  "markdown>=3.7,<4.0",
  "pymupdf>=1.27.2,<2.0.0",
  "python-docx>=1.1.2,<2.0.0",
  "pillow>=12.2.0,<13.0.0",
  "rapidocr_onnxruntime>=1.4.4,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.3.4,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "ruff>=0.8.4,<1.0.0",
]

[build-system]
requires = ["setuptools>=75.6.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
